Build and compile NIMBLE model and MCMC once This function is exported for use in 'future' and is not meant to be called by user.

Build and compile NIMBLE model and MCMC once This function is exported for use in 'future' and is not meant to be called by user.

Usage

build_ivd_model(code, constants, dummy_data, dummy_inits, useWAIC = TRUE)

Arguments

code

Nimble code

constants

Constants

dummy_data

Data

dummy_inits

inits

useWAIC

Defaults to TRUE. Nimble argument

Value

A named list with two elements:

cmodel: The compiled NIMBLE model object produced by compileNimble().

cmcmc: The compiled NIMBLE MCMC object, created using buildMCMC() and compileNimble(), configured to monitor the model parameters (including WAIC monitors if useWAIC = TRUE).

The function is intended for internal use (e.g., within parallel workers) and is not meant to be called directly by end users.